backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
python-magic-bin>=0.4,<1.0
dj-database-url>=2.0,<3.0
gunicorn>=21.0,<23.0
whitenoise>=6.5,<7.0
Pillow>=10.0,<12.0
